shell provider will reset wordpress settings after import wpcli will not install if already on machine

+++ scripts/install-wpcli.sh
@@ -1,7 +1,10 @@
 #!/usr/bin/env bash
 
+if [! -f /usr/local/bin/wp]
+then
 echo "Installing WP-CLI"
 curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
 sudo chmod +x wp-cli.phar
 sudo mv wp-cli.phar /usr/local/bin/wp
 echo "WP-CLI Installed"
+fi
